% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcg_cost}
\alias{bcg_cost}
\title{Costs in transition matrix format}
\format{
A 7-by-15 matrix
\describe{
  \item{rows}{from state}
  \item{columns}{to state}
  ...
}
}
\usage{
bcg_cost
}
\description{
Costs in transition matrix format
}
\keyword{datasets}
