% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcg_probs}
\alias{bcg_probs}
\title{Probabilities in transition matrix format}
\format{
A 7-by-15 matrix
\describe{
  \item{rows}{from state}
  \item{columns}{to state}
  ...
}
}
\usage{
bcg_probs
}
\description{
Probabilities in transition matrix format
}
\keyword{datasets}
