% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_joint_probs.R
\name{branch_joint_probs.transmat}
\alias{branch_joint_probs.transmat}
\title{branch_joint_probs.transmat}
\usage{
\method{branch_joint_probs}{transmat}(model, nodes = NA, ...)
}
\description{
branch_joint_probs.transmat
}
