% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEdecisiontree_recursive.R
\name{CEdecisiontree_recursive}
\alias{CEdecisiontree_recursive}
\title{CEdecisiontree_recursive}
\usage{
CEdecisiontree_recursive(tree, vals)
}
\arguments{
\item{tree}{List of children by parents}

\item{vals}{Node labels, probabililities and costs; dataframe}
}
\value{
Expected cost at root node
}
\description{
CEdecisiontree_recursive
}
\examples{

tree <-
  list("1" = c(2,3),
       "2" =  c(4,5),
       "3" =  c(6,7),
       "4" =  c(),
       "5" =  c(),
       "6" =  c(),
       "7" =  c())

vals <-
  data.frame(node = 1:7,
             probs = c(NA, rep(0.5, 6)),
             cost = c(10,2,3,16,5,6,7))

 CEdecisiontree_recursive(tree, vals)

}
\seealso{
CEdecisiontree
}
