% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEdecisiontree_recursive.R
\name{CEdecisiontree_recursive}
\alias{CEdecisiontree_recursive}
\title{Cost-effectiveness decision tree using recursive approach}
\usage{
CEdecisiontree_recursive(node, tree, dat)
}
\arguments{
\item{node}{Node at which total expected value is to be calculate at}

\item{tree}{List of children by parents}

\item{vals}{Node labels, probabililities and costs; dataframe}
}
\value{
Expected cost at root node
}
\description{
Cost-effectiveness decision tree using recursive approach
}
\examples{

tree <-
  list("1" = c(2,3),
       "2" =  c(4,5),
       "3" =  c(6,7),
       "4" =  c(),
       "5" =  c(),
       "6" =  c(),
       "7" =  c())
dat <-
  data.frame(node = 1:7,
             prob = c(NA, rep(0.5, 6)),
             vals = c(10,2,3,16,5,6,7))

root <- names(tree)[1]
CEdecisiontree_recursive(node = root, tree, dat)

}
\seealso{
CEdecisiontree
}
\concept{CEdecisiontree}
