% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ce_tree_df.R
\name{create_ce_tree_df}
\alias{create_ce_tree_df}
\title{Create cost-effectiveness decision tree data frame input}
\usage{
create_ce_tree_df(label_branch_tab, label_val_tab, tree_struc)
}
\arguments{
\item{label_branch_tab}{Look up table}

\item{label_val_tab}{Look up table}

\item{tree_struc}{List of parent child branches}
}
\value{
tibble in long format
}
\description{
Using look up table, rather than separate lists,
of values and labels.
Can prepare these external to R and read them in.
}
