% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_psa_inputs.R
\name{create_psa_inputs}
\alias{create_psa_inputs}
\title{Create PSA inputs}
\usage{
create_psa_inputs(tree_dat, n = 1000)
}
\arguments{
\item{tree_dat}{parent-child and data format but with distributions}

\item{n}{Number of samples; default 1000}
}
\value{
tree_dat realisations
}
\description{
Create PSA inputs
}
