% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree.R
\name{dectree}
\alias{dectree}
\title{Decision Tree}
\usage{
dectree(tree_dat, label_probs_distns, label_costs_distns, state_list, n = 100)
}
\arguments{
\item{tree_dat}{decision tree data}

\item{label_probs_distns}{probability distribution names}

\item{label_costs_distns}{cost distribution names}

\item{state_list}{State list}

\item{n}{default 100}
}
\value{

}
\description{
Decision Tree
}
\examples{

}
