% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_recursive2.R
\name{dectree_expected_recursive2}
\alias{dectree_expected_recursive2}
\title{Cost-effectiveness decision tree using recursive approach for more than 2 child nodes}
\usage{
dectree_expected_recursive2(node, tree, dat)
}
\arguments{
\item{node}{Node at which total expected value is to be calculate at}

\item{tree}{List of children by parents}

\item{dat}{Node labels, branch probabilities and value; data frame}
}
\value{
Expected value at root node
}
\description{
Cost-effectiveness decision tree using recursive approach for more than 2 child nodes
}
\examples{

tree <-
  list("1" = c(2,3,5),
       "2" =  4,
       "3" =  c(6,7),
       "4" =  c(),
       "5" =  c(),
       "6" =  c(),
       "7" =  c())
dat <-
  data.frame(node = 1:7,
             prob = c(NA, rep(1, 6)),
             vals = c(10,2,3,16,5,6,7))

root <- names(tree)[1]
dectree_expected_recursive2(node = root, tree, dat)

}
\seealso{
CEdecisiontree

Other CEdecisiontree:
\code{\link[=dectree_expected_recursive]{dectree_expected_recursive()}}
}
\concept{CEdecisiontree}
