% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_S3.R
\name{dectree_expected_values.dat_long}
\alias{dectree_expected_values.dat_long}
\title{dectree_expected_values.dat_long}
\usage{
\method{dectree_expected_values}{dat_long}(model)
}
\description{
dectree_expected_values.dat_long
}
