% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values.R,
%   R/dectree_expected_values_S3.R
\name{dectree_expected_values.default}
\alias{dectree_expected_values.default}
\alias{dectree_expected_values}
\title{Decision Tree Expected Values}
\usage{
\method{dectree_expected_values}{default}(vals, p, dat = NA)

dectree_expected_values(model, ...)
}
\arguments{
\item{vals}{Values on each edge/branch e.g. costs or QALYs (array)}

\item{p}{Transition probabilities matrix}

\item{dat}{Long node-edge value array; default: \code{NA}}

\item{model}{List as \code{define_model()} output of type \code{tree_dat}, \code{transmat} or \code{dat_long}}
}
\value{
Expected value at each node (vector)
}
\description{
Root node expected value as the weighted mean of
probability and edge/node values e.g. costs or QALYS.
}
\details{
The expected value at each node is calculate by

\deqn{\hat{C}_i = C_i + \sum p_{ij} \hat{C}_j}

The default calculation assumes that the costs are associated with the nodes.
An alternative would be to associated them with the edges.
For total expected cost this doesnt matter but for
the other nodes this is different to assuming the
costs are assigned to the nodes. The expected value would then be

\deqn{\hat{C}_i = \sum p_{ij} (C_{ij} + \hat{C}_j)}
}
\examples{

}
