% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_S3.R
\name{dectree_expected_values}
\alias{dectree_expected_values}
\alias{dectree_expected_values.tree_dat}
\alias{dectree_expected_values.transmat}
\alias{dectree_expected_values.dat_long}
\title{Cost-effectiveness decision tree expected values}
\usage{
dectree_expected_values(model)

\method{dectree_expected_values}{tree_dat}(model)

\method{dectree_expected_values}{transmat}(model)

\method{dectree_expected_values}{dat_long}(model)
}
\arguments{
\item{model}{Object of \code{define_model()} output of type
\code{tree_dat}, \code{transmat} or \code{dat_long}}
}
\value{
Expected value at each node
}
\description{
Root node expected value as the weighted mean of
probability and edge/node values e.g. costs or QALYS.
}
\details{
The expected value at each node is calculate by

\deqn{\hat{c}_i = c_i + \sum p_{ij} \hat{c}_j}

The default calculation assumes that the costs are associated with the nodes.
An alternative would be to associate them with the edges.
For total expected cost this doesn't matter but for
the other nodes this is different to assuming the
costs are assigned to the nodes. The expected value would then be

\deqn{\hat{c}_i = \sum p_{ij} (c_{ij} + \hat{c}_j)}
}
\examples{

}
\seealso{
define_model
}
