% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_S3.R
\name{dectree_expected_values.transmat}
\alias{dectree_expected_values.transmat}
\title{dectree_expected_values.transmat}
\usage{
\method{dectree_expected_values}{transmat}(model)
}
\description{
dectree_expected_values.transmat
}
