% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_S3.R
\name{dectree_expected_values.tree_dat}
\alias{dectree_expected_values.tree_dat}
\title{dectree_expected_values.tree_dat}
\usage{
\method{dectree_expected_values}{tree_dat}(model)
}
\description{
dectree_expected_values.tree_dat
}
