% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_node.R
\name{dectree_expected_values_node}
\alias{dectree_expected_values_node}
\title{Decision Tree Expected Values With Node Costs}
\usage{
dectree_expected_values_node(vals, p)
}
\arguments{
\item{vals}{Values on each edge/branch e.g. costs or QALYs (array)}

\item{p}{Transition probabilities (array)}
}
\value{
expected value at each node (list)
}
\description{
\deqn{\hat{C}_i = C_i + \sum p_{ij} \hat{C}_j}
}
