% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{f_NodeUniform}
\alias{f_NodeUniform}
\title{f_NodeUniform}
\usage{
f_NodeUniform(f)
}
\arguments{
\item{f}{function name}
}
\description{
TODO: this is probably better than repeating
most of the function body for each.
}
\examples{


meanNodeUnif <- f_NodeUniform(mean_distributions)
meanNodeUnif(unif_distn)

sampleNodeUnif <- f_NodeUniform(sample_distributions)
sampleNodeUnif(unif_distn)

}
\seealso{
\link{mean_distributions} \link{sample_distributions}
}
