% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_complementary_probs.R
\name{fill_complementary_probs}
\alias{fill_complementary_probs}
\title{Fill complementary probabilities}
\usage{
fill_complementary_probs(probs_names)
}
\arguments{
\item{probs_names}{long format tree object}
}
\value{
dataframe
}
\description{
Only one of each pair of branches is assigned a probability
and then the other event probability is filled-in afterwards.
This is good because specify fewer input values and
if sampling probabilities we don't know the complementary
probability.
}
\examples{

}
