% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_children_list.R
\name{get_children_list}
\alias{get_children_list}
\title{Get tree children list}
\usage{
get_children_list(transmat)
}
\arguments{
\item{transmat}{from-to matrix with \code{NA} for missing values.}
}
\value{
list
}
\description{
Get tree children list by parents from a transition matrix.
}
\examples{

}
