% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{get_sd_from_normalCI}
\alias{get_sd_from_normalCI}
\title{Get Standard Deviation from Normal Confidence Interval}
\usage{
get_sd_from_normalCI(n, x_bar = NA, upperCI = NA, lowerCI = NA)
}
\arguments{
\item{n}{Sample size}

\item{x_bar}{Mean}

\item{upperCI}{Upper 95% CI}

\item{lowerCI}{Lower 95% CI}
}
\value{
sd
}
\description{
Formula taken from
\href{http://stats.stackexchange.com/questions/30402/how-to-calculate-mean-and-standard-deviation-in-r-given-confidence-interval-and}{Stack Exchange}
}
