% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_to_transmat.R
\name{insert_to_transmat}
\alias{insert_to_transmat}
\alias{insert_to_costmat}
\alias{insert_to_probmat}
\title{Insert to transition matrix}
\usage{
insert_to_transmat(dat, mat)

insert_to_costmat(dat, mat)

insert_to_probmat(dat, mat)
}
\arguments{
\item{dat}{long from-to-vals array. Either cost or prob.}

\item{mat}{transition matrix}
}
\value{
matrix
}
\description{
Insert values into a transition matrix
from a long array of \code{from}, \code{to} values

Insert values into a cost matrix
from a long array of \code{from}, \code{to} values

Insert values into a probability matrix
from a long array of \code{from}, \code{to} values
}
\examples{

#insert_to_transmat(dat, mat)

}
