% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_prob_matrix}
\alias{is_prob_matrix}
\title{Is object a transition probability matrix?}
\usage{
is_prob_matrix(probs)
}
\arguments{
\item{probs}{matrix}
}
\description{
Is object a transition probability matrix?
}
\examples{
\dontrun{
probs <- matrix(c(1,0,0,1), nrow = 2)
is_prob_matrix(probs)

probs <- matrix(c(2,0,-1,1), nrow = 2)
assert_that(is_prob_matrix(probs))
}
}
\concept{assertions}
