% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_xx_transmat.R
\name{long_to_transmat}
\alias{long_to_transmat}
\title{Long format to transition matrix}
\usage{
long_to_transmat(dat)
}
\arguments{
\item{dat}{array of from, to, prob, vals}
}
\value{
transition matrix
}
\description{
Long format to transition matrix
}
\examples{

}
