% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchlabel_prob_fns.R
\name{match_branchlabel_to_prob}
\alias{match_branchlabel_to_prob}
\title{Match branch label to probabilities}
\usage{
match_branchlabel_to_prob(probs_names, branch_probs_long)
}
\arguments{
\item{branch_probs_long}{}
}
\value{
dataframe
}
\description{
Match branch label to probabilities
}
