% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{meanNodeUniform}
\alias{meanNodeUniform}
\title{Mean of a data.tree uniform node}
\usage{
meanNodeUniform(node)
}
\arguments{
\item{node}{data.tree node}
}
\description{
Mean of a data.tree uniform node
}
\seealso{
\link{mean_distributions}
}
