% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{MoM_beta}
\alias{MoM_beta}
\title{Method of Moments Beta Distribution Parameter Transformation}
\usage{
MoM_beta(xbar, vbar)
}
\arguments{
\item{xbar}{Mean}

\item{vbar}{Variance}
}
\value{
a and b of Beta(a,b)
}
\description{
Could alternatively use the beta-PERT \code{\link{rpert}} with maximum and minimum instead of variance.
}
\seealso{
\link{rpert}
}
