% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{MoM_gamma}
\alias{MoM_gamma}
\title{Method of Moments Gamma Distribution Parameter Transformation}
\usage{
MoM_gamma(mean, var)
}
\arguments{
\item{mean}{Mean}

\item{var}{Variance}
}
\value{
shape, scale
}
\description{
Method of Moments Gamma Distribution Parameter Transformation
}
\seealso{
\link{MoM_beta}
}
