% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xxx.R
\name{new_dat_long}
\alias{new_dat_long}
\title{new_dat_long}
\usage{
new_dat_long(dat_long, fill_edges = TRUE, fill_probs = FALSE)
}
\arguments{
\item{dat_long}{Long format tree data}

\item{fill_edges}{If need missing edges to connect to a sink state}

\item{fill_probs}{Fill in missing probabilities}
}
\description{
new_dat_long
}
