% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xxx.R
\name{new_models}
\alias{new_models}
\alias{new_transmat}
\alias{new_tree_dat}
\alias{new_dat_long}
\title{New Model Constructors}
\usage{
new_transmat(transmat, ...)

new_tree_dat(tree_dat, ...)

new_dat_long(dat_long, fill_edges = TRUE, fill_probs = FALSE)
}
\arguments{
\item{transmat}{Transition probability matrix. Rows are from nodes and columns are to nodes.}

\item{...}{Additional arguments}

\item{tree_dat}{Hierarchical tree structure of parents and children in a list of vectors with integer values.}

\item{dat_long}{Long format data frame with from, to, prob, vals columns.}

\item{fill_edges}{If need missing edges to connect to a sink state; logical}

\item{fill_probs}{Fill in missing probabilities; logical}
}
\description{
New Model Constructors

New transition matrix

New tree data

New long data
}
