% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probs}
\alias{probs}
\title{Probabilities in transition matrix format}
\format{
A 3-by-7 matrix
\describe{
  \item{rows}{from state}
  \item{columns}{to state}
  ...
}
}
\usage{
probs
}
\description{
Probabilities in transition matrix format
}
\keyword{datasets}
