% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{rbeta_more_params}
\alias{rbeta_more_params}
\title{Helper function for Beta distribution sampling with alternative parameters}
\usage{
rbeta_more_params(
  mean = NA,
  sd = NA,
  a = NA,
  b = NA,
  shape1 = NA,
  shape2 = NA,
  ...
)
}
\description{
Helper function for Beta distribution sampling with alternative parameters
}
