% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{rgamma_more_params}
\alias{rgamma_more_params}
\title{Helper function for Gamma distribution sampling with alternative parameters}
\usage{
rgamma_more_params(
  mean = NA,
  sd = NA,
  var = NA,
  shape = NA,
  scale = 1,
  rate = NA,
  ...
)
}
\description{
Helper function for Gamma distribution sampling with alternative parameters
}
