% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{rpert}
\alias{rpert}
\title{Sample from Beta-PERT Distribution}
\usage{
rpert(n, x.min, x.max, x.mode, lambda = 4)
}
\arguments{
\item{n}{Sample size}

\item{x.min}{Lower limit}

\item{x.max}{Upper limit}

\item{x.mode}{Mode}

\item{lambda}{Shape parameter}
}
\value{
sampled value
}
\description{
see https://reference.wolfram.com/language/ref/PERTDistribution.html
}
