% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cedectree.R
\name{run_cedectree}
\alias{run_cedectree}
\title{Run Cost-effectiveness Decision Tree}
\usage{
run_cedectree(
  tree_dat,
  label_probs_distns = NULL,
  label_costs_distns = NULL,
  label_health_distns = NULL,
  state_list = NULL,
  n = 100
)
}
\arguments{
\item{tree_dat}{Decision tree data in `dat_long` format}

\item{label_probs_distns}{Probability distribution names}

\item{label_costs_distns}{Cost distribution names}

\item{label_health_distns}{Health value distribution names}

\item{state_list}{State list sets, usually terminal nodes}

\item{n}{Number of PSA samples; default 100}
}
\value{
List of cost, health `dectree()` output
}
\description{
Wrapper for `dectree()` for both costs and health value.
}
\seealso{
dectree
}
