% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{sample_distributions}
\alias{sample_distributions}
\title{Sample from Standard Distributions}
\usage{
sample_distributions(param.distns)
}
\arguments{
\item{param.distns}{List of distribution names and their respective parameter values}
}
\value{
vector of sample points
}
\description{
Supply a list of defined distributions (log-normal, beta, gamma, uniform, triangle)
and one realisation is taken of each.
}
\examples{

sample_distributions(param.distns = list(distn = "unif", params = c(min=0, max=1)))
sample_distributions(param.distns = list(distn = "lognormal", params = c(mean=10, sd=1)))
sample_distributions(param.distns = list(distn = "beta", params = c(mean=0.1, sd=0.1)))
sample_distributions(param.distns = list(distn = "beta", params = c(a=0.1, b=0.1)))
sample_distributions(param.distns = list(list(distn = "beta", params = c(a=0.1, b=0.1)),
                                         list(distn = "beta", params = c(a=0.1, b=0.1))))

}
