% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{sampleNode}
\alias{sampleNode}
\title{Sample a data.tree Node}
\usage{
sampleNode(node)
}
\arguments{
\item{node}{data.tree node}
}
\description{
TODO: could use NextMethod()
i.e. write sample_distribution.node function
}
\examples{

osNode
osNode$Get(sampleNode)

}
\seealso{
\link{sample_distributions}
}
