% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{sampleNodeUniform}
\alias{sampleNodeUniform}
\title{Sample a data.tree uniform node}
\usage{
sampleNodeUniform(node)
}
\arguments{
\item{node}{data.tree node}
}
\description{
Sample a data.tree uniform node
}
\examples{

rprob <- osNode$Get(sampleNodeUniform)

}
\seealso{
\link{sample_distributions}
}
