% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_pop.R
\name{terminal_pop}
\alias{terminal_pop}
\title{Terminal Leaf Node Populations}
\usage{
terminal_pop(model, state_list)
}
\arguments{
\item{model}{dat_long format}

\item{state_list}{Groups of (usually) terminal nodes; List of vectors}
}
\value{
Vector of probabilities
}
\description{
Joint probabilities of sets of nodes.
This is useful to then provide starting state
populations to a Markov model.
}
\examples{

tree_dat <-
 tibble::tribble(
   ~from, ~to, ~vals, ~prob,
   1,  2,   10,   0.7,
   1,  3,   NA,   0.3,
   2,  4,  100,   0.1,
   2,  5,   NA,   0.9,
   3,  6,  100,   0.9,
   3,  7,   NA,   0.1)

term_pop <-
  define_model(dat_long = tree_dat) |>
  terminal_pop(state_list = c(4,5,6,7))

sum(unlist(term_pop))

}
