% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmat_to_child_list.R
\name{transmat_to_child_list}
\alias{transmat_to_child_list}
\title{Create tree children list from transition matrix}
\usage{
transmat_to_child_list(transmat)
}
\arguments{
\item{transmat}{from-to matrix with \code{NA} for missing values.}
}
\value{
list
}
\description{
Create tree children list by parents from a transition matrix.
}
\examples{

transmat <-
  list(prob = matrix(data = c(NA, 0.5, 0.5), nrow = 1),
         vals = matrix(data = c(NA, 1, 2), nrow = 1)
       )
transmat_to_child_list(transmat$prob)

}
