% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_xx_transmat.R
\name{transmat_to_long}
\alias{transmat_to_long}
\title{Transition matrix to long format}
\usage{
transmat_to_long(probs)
}
\arguments{
\item{probs}{probability transition matrix}
}
\value{
array of from, to, prob
}
\description{
Transition matrix to long format
}
\examples{

}
