% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{validation}
\alias{validation}
\alias{validate_transmat}
\alias{validate_tree_dat}
\alias{validate_dat_long}
\title{New Model Validation}
\usage{
validate_transmat(transmat)

validate_tree_dat(tree_dat)

validate_dat_long(dat_long)
}
\arguments{
\item{transmat}{Transition probability matrix. Rows are from nodes and columns are to nodes.}

\item{tree_dat}{Hierarchical tree structure of parents and children in a list of vectors with integer values.}

\item{dat_long}{Long format data frame with from, to, prob, vals columns.}
}
\value{
transmat

tree_dat

dat_long
}
\description{
New Model Validation
}
