% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_joint_probs.R
\name{branch_joint_probs.dat_long}
\alias{branch_joint_probs.dat_long}
\title{branch_joint_probs.dat_long}
\usage{
\method{branch_joint_probs}{dat_long}(model, nodes, ...)
}
\description{
branch_joint_probs.dat_long
}
