% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_joint_probs.R
\name{branch_joint_probs}
\alias{branch_joint_probs}
\title{Branch Joint Probabilities}
\usage{
branch_joint_probs(model, ...)
}
\description{
Provides a measure of the chances of following
particular paths.
}
\details{
These probabilities could be used to weight branch costs
or QALYs to indicate the relative contribution to the
total expected value.
}
