% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_joint_probs.R
\name{branch_joint_probs}
\alias{branch_joint_probs}
\title{Branch Joint Probability}
\usage{
branch_joint_probs(probs)
}
\arguments{
\item{probs}{Branch conditional probabilities (array)}
}
\description{
Branch Joint Probability
}
\examples{
branch_joint_probs(probs) * cost

}
