% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_joint_probs.R
\name{branch_joint_probs.transmat}
\alias{branch_joint_probs.transmat}
\title{branch_joint_probs.transmat}
\usage{
\method{branch_joint_probs}{transmat}(model, nodes = NA)
}
\arguments{
\item{model}{Branch conditional probabilities (matrix)}

\item{nodes}{Which nodes to return; default to all}
}
\value{
Transition matrix with joint probabilities
}
\description{
branch_joint_probs.transmat
}
\examples{

model <-
  define_model(
    transmat =
      list(prob =
             matrix(data = c(NA, 0.5, 0.5, NA,  NA,  NA,  NA,
                             NA, NA, NA,   0.1, 0.9, NA,  NA,
                             NA, NA, NA,   NA,  NA,  0.9, 0.1),
                    nrow = 3,
                    byrow = TRUE),
           vals =
             matrix(data = c(NA, 1,  5,  NA, NA, NA, NA,
                             NA, NA, NA, 1,  9,  NA, NA,
                             NA, NA, NA, NA, NA, 9,  1),
                    nrow = 3,
                    byrow = TRUE)))
model

branch_joint_probs(model)
branch_joint_probs(model)*model$vals

}
