% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_list_to_transmat.R
\name{child_list_to_transmat}
\alias{child_list_to_transmat}
\title{Create transition matrix from tree children list}
\usage{
child_list_to_transmat(tree)
}
\arguments{
\item{tree}{parent-child node list.}
}
\value{
matrix
}
\description{
Create transition matrix from tree children list by parents.
}
\examples{

child_list_to_transmat(tree =
                         list('1' = c(2, 3),
                              '2' = c(4),
                              '3' = c(6, 7)))

}
