% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree.R
\name{dectree}
\alias{dectree}
\title{Decision Tree}
\usage{
dectree(
  tree_dat,
  label_probs_distns = NULL,
  label_vals_distns = NULL,
  state_list = NULL,
  n = 100
)
}
\arguments{
\item{tree_dat}{Decision tree data in `dat_long` format}

\item{label_probs_distns}{Probability distribution names}

\item{label_vals_distns}{Value distribution names}

\item{state_list}{State list sets, usually terminal nodes}

\item{n}{Number of PSA samples; default 100}
}
\value{
List of expected values at each node,
        joint probabilities at terminal state set
        and PSA samples of these if distributions provided.
}
\description{
Single edge value (e.g. cost or QALY) wrapper.
}
