% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values.R,
%   R/dectree_expected_values_S3.R
\name{dectree_expected_values.default}
\alias{dectree_expected_values.default}
\alias{dectree_expected_values}
\title{Decision Tree Expected Values}
\usage{
\method{dectree_expected_values}{default}(vals, p, dat = NA)

dectree_expected_values(model, ...)
}
\arguments{
\item{vals}{Values on each edge/branch e.g. costs or QALYs (array)}

\item{p}{Transition probabilities (array)}

\item{dat}{default: NA}

\item{model}{\code{define_model} output of type \code{tree_dat}, \code{transmat} or \code{dat_long}}
}
\value{
expected value at each node (list)
}
\description{
Root node expected value as the weighted mean of
probability and edge/node values e.g. costs or QALYS.
}
\details{
At the moment, the default calculation assumes
that the costs are associated with the edges and not
the 'to' node.
For total expected cost this doesnt matter but for
the other nodes this is different to assuming the
costs are assigned to the nodes.

\deqn{C_i = \sum p_{ij} (C_{ij} + C_j)}
}
\examples{

}
