% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values.R
\name{dectree_expected_values}
\alias{dectree_expected_values}
\title{Decision Tree Expected Values}
\usage{
dectree_expected_values(vals, p)
}
\arguments{
\item{vals}{Values on each edge/branch e.g. costs or QALYs (array)}

\item{p}{Transition probabilities (array)}
}
\value{
expected value at each node (list)
}
\description{
\deqn{C_i = \sum p_{ij} (C_{ij} + C_j)}
}
\examples{
dectree_expected_values(cost, probs)

}
