% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree_expected_values_S3.R
\name{dectree_expected_values.tree_dat}
\alias{dectree_expected_values.tree_dat}
\alias{dectree_expected_values.transmat}
\alias{dectree_expected_values.dat_long}
\title{dectree_expected_values.tree_dat}
\usage{
\method{dectree_expected_values}{tree_dat}(model)

\method{dectree_expected_values}{transmat}(model)

\method{dectree_expected_values}{dat_long}(model)
}
\description{
dectree_expected_values.tree_dat

dectree_expected_values.transmat

dectree_expected_values.dat_long
}
