% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R
\name{define_model}
\alias{define_model}
\title{Define model}
\usage{
define_model(transmat, tree_dat, dat_long)
}
\arguments{
\item{transmat}{Transition probability matrix (from-to node)}

\item{tree_dat}{Hierarchical tree structure of parents and children}

\item{dat_long}{Long dataframe with from, to, prob, vals columns}
}
\value{
transmat, tree_dat or dat_long
}
\description{
Define model
}
\examples{

define_model(transmat =
              list(prob = matrix(data=c(NA, 0.5, 0.5), nrow = 1),
                   vals = matrix(data=c(NA, 1, 2), nrow = 1)
              ))

define_model(tree_dat =
              list(child = list("1" = c(2, 3),
                                "2" = NULL,
                                "3" = NULL),
                   dat = data.frame(node = 1:3,
                                    prob = c(NA, 0.5, 0.5),
                                    vals = c(0, 1, 2))
              ))

define_model(dat_long = data.frame(from = c(NA, 1, 1),
                                   to = 1:3,
                                   prob = c(NA, 0.5, 0.5),
                                   vals = c(0, 1, 2)))
}
