% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_complementary_probs.R
\name{fill_complementary_probs}
\alias{fill_complementary_probs}
\title{Fill Complementary Probabilities}
\usage{
fill_complementary_probs(dat_tree)
}
\arguments{
\item{dat_tree}{Long format tree object}
}
\value{
Long format tree object
}
\description{
Only one of each pair of branches is assigned a probability
and then the other event probability is filled-in afterwards.
This is good because specify fewer input values and
if sampling probabilities we don't know the complementary
probability.
}
\details{
Only works for binary trees or when one of n is missing.
}
\examples{
dat <- tibble::tribble(~from, ~to, ~prob, ~vals,
                       1,  2,  NA,    10,
                       1,  3,  0.8,    1,
                       2,  4,  0.2,   10,
                       2,  5,  NA,     1,
                       3,  6,  0.2,   10,
                       3,  7,  0.8,    1)
fill_complementary_probs(dat)

}
