% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchlabel_prob_fns.R
\name{fill_complementary_probs}
\alias{fill_complementary_probs}
\title{Fill complementary probabilities}
\usage{
fill_complementary_probs(probs_names)
}
\arguments{
\item{probs_names}{}
}
\value{
dataframe
}
\description{
Fill complementary probabilities
}
