% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_children_list.R
\name{get_children_list}
\alias{get_children_list}
\title{get_children_list}
\usage{
get_children_list(transmat)
}
\arguments{
\item{transmat}{from-to matrix with NA for missing values.}
}
\description{
get_children_list
}
