% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_xx_transmat.R
\name{long_to_transmat}
\alias{long_to_transmat}
\title{Long format to transition matrix}
\usage{
long_to_transmat(dat)
}
\arguments{
\item{dat}{array of from, to, prob, vals}
}
\value{
transition matrix
}
\description{
Long format to transition matrix
}
\examples{

data <- data.frame(from = c(NA,1, 1),
                  to = c(1, 2, 3),
                  prob = c(NA, 0.5, 0.5),
                  vals = c(0, 1, 2))
long_to_transmat(dat)

}
