% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_branch_to_label.R
\name{match_branch_to_label}
\alias{match_branch_to_label}
\title{Match branch to label}
\usage{
match_branch_to_label(probs_long, probs_from_to_lookup)
}
\arguments{
\item{probs_long}{Long format array tree object}

\item{probs_from_to_lookup}{edge-label look-up table}
}
\value{
dataframe
}
\description{
Assume that there are node labels and these may be non-unique.
We assign these labels by joining a lookup table of labels and edges
with the tree object, in this case the long array format.
}
\details{
Separating the tree structure and labelling means that we can
reuse the same tree with different labels
e.g. another test or treatment
}
\examples{
probs_long <-
  tibble::tribble(~from, ~to, ~prob,
                  1, 2, 0.5,
                  1, 3, 0.5)
pname_from_to <-
  tibble::tribble(~from, ~to, ~name,
                  1, 2, "pos",
                  1, 3, "neg")

match_branch_to_label(probs_long, pname_from_to)

}
\seealso{
\link{match_branchlabel_to_prob}
}
