% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_branchlabel_to_prob.R
\name{match_branchlabel_to_prob}
\alias{match_branchlabel_to_prob}
\title{Match branch label to probabilities}
\usage{
match_branchlabel_to_prob(probs_names, branch_probs_long)
}
\arguments{
\item{probs_names}{label-probability look-up}

\item{branch_probs_long}{long format tree object with labels}
}
\value{
dataframe
}
\description{
We have a look-up table of labels and values.
We can join the values to nodes/edges via their labels.
}
\examples{

}
\seealso{
\link{match_branch_to_label}
}
