% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xxx.R
\name{new_models}
\alias{new_models}
\alias{new_transmat}
\alias{new_tree_dat}
\alias{new_dat_long}
\title{New Model Constructors}
\usage{
new_transmat(transmat, ...)

new_tree_dat(tree_dat, ...)

new_dat_long(dat_long, fill_edges = TRUE, fill_probs = FALSE)
}
\arguments{
\item{dat_long}{Long format tree data}

\item{fill_edges}{If need missing edges to connect to a sink state; logical}

\item{fill_probs}{Fill in missing probabilities; logical}
}
\description{
New Model Constructors

new_transmat

new_tree_dat

new_dat_long
}
