% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_pop.R
\name{terminal_pop}
\alias{terminal_pop}
\title{Terminal Leaf Node Populations}
\usage{
terminal_pop(model, state_list)
}
\arguments{
\item{model}{dat_long format}

\item{state_list}{Groups of usually terminal nodes}
}
\value{
Vector of probabilities
}
\description{
Joint probabilities of sets of nodes.
This is useful to then provide starting state
populations to a Markov model.
}
\examples{

}
