% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_pop.R
\name{terminal_pop}
\alias{terminal_pop}
\title{terminal_pop}
\usage{
terminal_pop(model, state_list)
}
\arguments{
\item{model}{long format}

\item{state_list}{}
}
\value{

}
\description{
terminal_pop
}
\examples{

}
