% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_binarytree.R
\name{trans_binarytree}
\alias{trans_binarytree}
\title{trans_binarytree}
\usage{
trans_binarytree(names, depth = 2)
}
\arguments{
\item{depth}{}
}
\description{
This is adapted from mstate::trans.illness
Create a complete binary tree transition matrix.
}
