% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmat_to_child_list.R
\name{transmat_to_child_list}
\alias{transmat_to_child_list}
\title{Get tree children list from transition matrix}
\usage{
transmat_to_child_list(transmat)
}
\arguments{
\item{transmat}{from-to matrix with \code{NA} for missing values.}
}
\value{
list
}
\description{
Get tree children list by parents from a transition matrix.
}
\examples{

}
