% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{validation}
\alias{validation}
\alias{validate_transmat}
\alias{validate_tree_dat}
\alias{validate_dat_long}
\title{New Model Validation}
\usage{
validate_transmat(transmat)

validate_tree_dat(tree_dat)

validate_dat_long(dat_long)
}
\arguments{
\item{transmat}{Transition probability matrix}

\item{tree_dat}{Tree data}

\item{dat_long}{Long format data}
}
\description{
New Model Validation
}
